Warning Notification Component - initialized in App component uses Portal to render on root-warningNotification - state stored in context WarningNotificationContext (show: bool, message: string: type: string) - manages 3 types: SUCCESS, WARNING, ERROR
when dispatching using any of three types will render this compoenent and after 5 seconds it will unmount using settimeout

+++ client/src/components/BoardCard/BoardCard.js
@@ -44,12 +44,16 @@ const BoardCard = ({ boardInfo, boardId, isPinned, pinBoard, removeBoard, isAuth
 	};
 
 	const deleteBoardHandler = async () => {
-		const { error } = await fetchData({
-			method: "DELETE",
-			url: `/board/${boardId}`,
-			token: true,
-		});
-		if (!error) removeBoard(boardId);
+		const shouldDelete = window.confirm("are you sure you want to delete this board?")
+		if(shouldDelete){
+			const { error } = await fetchData({
+				method: "DELETE",
+				url: `/board/${boardId}`,
+				token: true,
+			});
+			if (!error) removeBoard(boardId);
+		} 
+		
 	};
 	const leavingEvent = async () => {
 		const { error } = await fetchData({
